backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication & Security
bcrypt==4.1.2
pyotp==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Client & Validation
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
python-magic==0.4.27
aiofiles==23.2.1

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
freezegun==1.2.2

# Security Testing
bandit==1.7.5
safety==2.3.5

# Code Quality
ruff==0.1.6
black==23.11.0
mypy==1.7.1